move domain usecases
