+++ features/home/src/main/java/com/rmakiyama/spatz/home/HomeViewModel.kt
@@ -7,8 +7,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.rmakiyama.spatz.domain.model.tweet.Tweet
 import com.rmakiyama.spatz.domain.result.Result
-import com.rmakiyama.spatz.domain.usecase.auth.LoadAuthUserUseCase
-import com.rmakiyama.spatz.domain.usecase.tweet.GetTweetsUseCase
+import com.rmakiyama.spatz.usecase.auth.LoadAuthUserUseCase
+import com.rmakiyama.spatz.usecase.tweet.GetTweetsUseCase
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import timber.log.Timber
