+++ usecase/src/main/java/com/rmakiyama/spatz/usecase/FlowUseCase.kt
@@ -1,4 +1,4 @@
-package com.rmakiyama.spatz.domain.usecase
+package com.rmakiyama.spatz.usecase
 
 import com.rmakiyama.spatz.domain.result.Result
 import kotlinx.coroutines.flow.Flow
