+++ usecase/src/main/java/com/rmakiyama/spatz/usecase/auth/SaveAuthUserUseCase.kt
@@ -1,13 +1,12 @@
-package com.rmakiyama.spatz.domain.usecase.auth
+package com.rmakiyama.spatz.usecase.auth
 
 import com.rmakiyama.spatz.domain.model.auth.AuthUser
 import com.rmakiyama.spatz.domain.repository.AuthRepository
-import com.rmakiyama.spatz.domain.usecase.SuspendUseCase
 import javax.inject.Inject
 
 class SaveAuthUserUseCase @Inject constructor(
     private val authRepository: AuthRepository
-) : SuspendUseCase<AuthUser, Unit>() {
+) : com.rmakiyama.spatz.usecase.SuspendUseCase<AuthUser, Unit>() {
 
     override suspend fun execute(command: AuthUser) {
         authRepository.save(command)
